it works!
I needed to understand how `?` worked a bit more and that
the failure::Error type worked with `?` but I'd still need to
return an `Ok(string)`. I get it now!

now with an invalid bit of json it does this the program prints:
```
error: key must be a string,
```

yay